refactor: split the code
